my_project/requirements.txt
Django==4.2
grpcio==1.58
grpcio-tools==1.58
uvicorn==0.22.0
